backend/requirements.txt
fastapi[all]
uvicorn
opencv-python-headless  # 無頭版，不需要GUI
mediapipe
numpy
python-dotenv